+++ instrumentation/github.com/aws/aws-sdk-go-v2/otelaws/attributes.go
@@ -16,6 +16,7 @@ package otelaws
 
 import "go.opentelemetry.io/otel/attribute"
 
+// AWS attributes.
 const (
 	OperationKey attribute.Key = "aws.operation"
 	RegionKey    attribute.Key = "aws.region"