+++ CHANGELOG.md
@@ -8,6 +8,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+### Fixed
+
+- Dockerfile based examples for `otelgin` and `otelmacaron`. (#767)
+
 ### Changed
 
 - Supported minimum version of Go bumped from 1.14 to 1.15. (#787)